requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework-simplejwt
djangorestframework
celery
redis
django-celery-beat
django-sms
django-sms[twilio]